add visual clue text
