+++ src/palette/elements/VisualClue/VisualClueDot.tsx
@@ -1,0 +1,67 @@
+import { Flex, useColor } from "palette"
+import React, { useEffect, useRef } from "react"
+import { Animated } from "react-native"
+
+export const VisualClueDot: React.FC = () => {
+  const size = 6
+  const duration = 1600
+
+  const scaleAnimation = useRef(new Animated.Value(1))
+  const opacityAnimation = useRef(new Animated.Value(1))
+
+  const color = useColor()
+
+  useEffect(() => {
+    Animated.loop(
+      Animated.sequence([
+        Animated.timing(scaleAnimation.current, {
+          toValue: 3,
+          duration,
+          useNativeDriver: true,
+        }),
+        Animated.timing(scaleAnimation.current, {
+          toValue: 1,
+          duration: 0,
+          useNativeDriver: true,
+        }),
+      ])
+    ).start()
+
+    Animated.loop(
+      Animated.sequence([
+        Animated.timing(opacityAnimation.current, {
+          toValue: 0.14,
+          duration: duration / 2,
+          useNativeDriver: true,
+        }),
+        Animated.timing(opacityAnimation.current, {
+          toValue: 0,
+          duration: duration / 2,
+          useNativeDriver: true,
+        }),
+      ])
+    ).start()
+  }, [])
+
+  return (
+    <Flex
+      style={{
+        height: size,
+        minWidth: size,
+        borderRadius: size / 2,
+        backgroundColor: color("blue100"),
+      }}
+    >
+      <Animated.View
+        style={{
+          height: size,
+          minWidth: size,
+          borderRadius: size / 2,
+          backgroundColor: color("blue100"),
+          transform: [{ scale: scaleAnimation.current }],
+          opacity: opacityAnimation.current,
+        }}
+      />
+    </Flex>
+  )
+}
