+++ src/palette/elements/VisualClue/index.tsx
@@ -1,0 +1,2 @@
+export * from "./VisualClueText"
+export * from "./VisualClueDot"
